tinydns: move functions into lib/tinydns
